streamlit-operator/requirements.txt
kopf[full-auth]~=1.36.0
kubernetes~=24.2.0
PyYAML~=6.0
pydantic~=2.12.5
urllib3<2.0.0
typing-extensions
